NavigationControllerImpl::GetNavigationApiHistoryEntryVectors() needs to handle a history nav without a pending entry
If a history traversal navigates multiple frames, it's possible to get a timing
where one completes before the last commits. In that case, when the last
commits, there will no longer be a pending entry even though we are committing
a history navigation.

GetNavigationApiHistoryEntryVectors() doesn't handle that case and will
incorrectly calculate the current index in the session history. This in turn
leads to incorrectly caulclating the NavigationApiHistoryEntryVectors.

Change-Id: I745c5849cbaf9457788e30bf07b8fbb99cfa47c0