[@scope] Support invalidation
This CL extends RuleFeatureSet to take into account scoped rules.

For example, for a rule:

  @scope (.a) {
   .b {}
  }

We need to add invalidation features equivalent to :is(.a .b, .a.b).
It is not enough to add features for just `.a .b`, because scopes
established by @scope rules *include* the scoping root, hence any match
status change for the from-selectors could potentially need to invalidate
the affected element itself---not just the descendants.

It is also possible to use the :scope pseudo-class to refer to the scoping
root itself. Currently, we don't extract any features from :scope, and
will therefore end up with whole-subtree invalidation in those cases.
We can improve this in the future.

Bug: 1280240
Change-Id: I158ca1a947233fdcd4f06609126371800cd8e63b