[LayoutNG] RequiresContentBeforeBreaking() and OOF fragmentation
We would previously always set RequiresContentBeforeBreaking() to true
for OOF positioned elements (see CL:3310887 for more details). However,
we should use the value of RequiresContentBeforeBreaking() of the CB as
the basis, instead.

Store an extra bit in NGContainingBlock to keep track of what
RequiresContentBeforeBreaking() should be for the CB. This will get
set when adjusting the newly created OOF fragmentainer descendant
in NGOutOfFlowLayoutPart. As a result, NGBoxFragmentBuilder::
AdjustOffsetsForFragmentainerDescendant() was renamed to
AdjustFragmentainerDescendant() since the method no longer is just
adjusting offsets.

tall-content-inside-constrained-block-003.tentative.html and
tall-content-inside-constrained-block-004.tentative.html were updated
based on the new logic, and two similar tests were added to
ensure this gets set correctly for |fixedpos_containing_block|.

Bug: 1319280
Change-Id: I450adb7e4a07c6948d76dcd3ec808a848483e8a2